pyproject.toml
[project]
name = "aai-capstone"
version = "0.1.0"
description = "Automated presentation generator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "colorama>=0.4.6",
    "python-pptx>=1.0.2",
    "pyyaml>=6.0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
